wip: put it all in the extension class
